Add maven pom.xml
This allows you to build the project by running
"mvn clean package". It will generate the "light" and
"all" version as well as the RequiredLibs. This also
would allow us to publish the artifact to the maven
repository.
Renamed folders to be lowercase. Uppercase in folder
names is not valid JAVA.